Port `load_kafka` and `save_kafka`
