+++ plugins/kafka/CMakeLists.txt
@@ -13,6 +13,7 @@ TenzirRegisterPlugin(
   TARGET kafka
   ENTRYPOINT src/plugin.cpp
   SOURCES GLOB "${CMAKE_CURRENT_SOURCE_DIR}/src/*.cpp"
+  BUILTINS GLOB "${CMAKE_CURRENT_SOURCE_DIR}/builtins/*.cpp"
   INCLUDE_DIRECTORIES include)
 
 find_package(RdKafka QUIET)
